refactor actionable event emails
