+++ tunga_tasks/utils.py
@@ -1,11 +1,16 @@
 import json
+
+from django.contrib.auth import get_user_model
+from django.db.models import When, Sum, Case, IntegerField, F
 from django.utils import six
 
 from allauth.socialaccount.providers.github.provider import GitHubProvider
 
+from tunga_auth.filterbackends import my_connections_q_filter
 from tunga_profiles.utils import get_app_integration
 from tunga_tasks.models import Integration, IntegrationMeta
-from tunga_utils.constants import APP_INTEGRATION_PROVIDER_SLACK, APP_INTEGRATION_PROVIDER_HARVEST, STATUS_ACCEPTED
+from tunga_utils.constants import APP_INTEGRATION_PROVIDER_SLACK, APP_INTEGRATION_PROVIDER_HARVEST, STATUS_ACCEPTED, \
+    USER_TYPE_DEVELOPER, VISIBILITY_MY_TEAM
 from tunga_utils.helpers import clean_meta_value, get_social_token, GenericObject
 
 
@@ -74,3 +79,61 @@ def save_integration_tokens(user, task_id, provider):
                 token_info.pop('token_secret')
                 token_info['refresh_token'] = app_integration.token_secret
     save_task_integration_meta(task_id, provider, token_info)
+
+
+def get_suggested_community_receivers(instance, user_type=USER_TYPE_DEVELOPER, respect_visibility=True):
+    # Filter users based on nature of work
+    queryset = get_user_model().objects.filter(
+        type=user_type
+    )
+
+    # Only developers on client's team
+    if respect_visibility and instance.visibility == VISIBILITY_MY_TEAM and user_type == USER_TYPE_DEVELOPER:
+        queryset = queryset.filter(
+            my_connections_q_filter(instance.user)
+        )
+
+    ordering = []
+
+    # Order by matching skills
+    task_skills = instance.skills.all()
+    if task_skills:
+        when = []
+        for skill in task_skills:
+            new_when = When(
+                userprofile__skills=skill,
+                then=1
+            )
+            when.append(new_when)
+        queryset = queryset.annotate(matches=Sum(
+            Case(
+                *when,
+                default=0,
+                output_field=IntegerField()
+            )
+        ))
+        ordering.append('-matches')
+
+    # Order developers by tasks completed
+    if user_type == USER_TYPE_DEVELOPER:
+        queryset = queryset.annotate(
+            tasks_completed=Sum(
+                Case(
+                    When(
+                        participation__task__closed=True,
+                        participation__user__id=F('id'),
+                        participation__status=STATUS_ACCEPTED,
+                        then=1
+                    ),
+                    default=0,
+                    output_field=IntegerField()
+                )
+            )
+        )
+        ordering.append('-tasks_completed')
+
+    if ordering:
+        queryset = queryset.order_by(*ordering)
+    if queryset:
+        return queryset[:15]
+    return
